+++ lib/workload/stateless/postgres_manager/function/alter-pg-db-owner.ts
@@ -1,11 +1,11 @@
 import { Client } from 'pg';
 import {
-  EventType,
   getMicroserviceConfig,
   getMicroserviceName,
   executeSqlWithLog,
   getRdsMasterSecret,
 } from './utils';
+import { EventType } from './type';
 
 export const handler = async (event: EventType) => {
   const microserviceConfig = getMicroserviceConfig();
