+++ lib/workload/stateless/postgres_manager/function/create-pg-db.ts
@@ -1,6 +1,6 @@
 import { Client } from 'pg';
+import { EventType } from './type';
 import {
-  EventType,
   getMicroserviceConfig,
   getMicroserviceName,
   executeSqlWithLog,
