+++ lib/workload/stateless/postgres_manager/function/utils.ts
@@ -1,22 +1,6 @@
 import { Client } from 'pg';
 import { SecretsManagerClient, GetSecretValueCommand } from '@aws-sdk/client-secrets-manager';
-
-/**
- * There are 2 ways of connecting from microservice to db
- */
-export enum DbAuthType {
-  RDS_IAM,
-  USERNAME_PASSWORD,
-}
-
-export type EventType = {
-  microserviceName: string;
-};
-
-export type MicroserviceConfig = {
-  name: string;
-  authType: DbAuthType;
-}[];
+import { MicroserviceConfig, EventType } from './type';
 
 /**
  * get microservice config from lambda environment
