+++ tsconfig.json
@@ -30,7 +30,6 @@
   "exclude": [
     "node_modules",
     "cdk.out",
-    "lib/workload/stateless/metadata_manager",
-    "lib/workload/stateless/postgres_manager"
+    "lib/workload/stateless/**",
   ]
 }
